delete-customer-cascade-01

+++ front-end/src/app/customer/customer.component.ts
@@ -3,19 +3,21 @@ import {HttpClient, HttpClientModule} from "@angular/common/http";
 import {NgForOf, NgIf} from "@angular/common";
 import {CustomerService} from "../services/customer.service";
 import {CustomerModel} from "../model/customer.model";
+import {FormsModule} from "@angular/forms";
 
 @Component({
   selector: 'app-customer',
   standalone: true,
   imports: [
-    NgIf,HttpClientModule, NgForOf
+    NgIf,HttpClientModule, NgForOf, FormsModule
   ],
   templateUrl: './customer.component.html',
   styleUrl: './customer.component.css'
 })
 export class CustomerComponent implements OnInit{
 
-  customers : Array<CustomerModel> = []
+  public customers : Array<CustomerModel> = []
+  public keyword: string = "";
 
   constructor(private customerService : CustomerService) {
   }
@@ -36,14 +38,26 @@ export class CustomerComponent implements OnInit{
   }
 
   deleteCustomer(customerId : number) {
+    if (confirm("You're About to Delete a Customer "))
     this.customerService.deleteCustomer(customerId)
       .subscribe({
         next : value => {
-          alert("Client Supprimé avec Succès !!!")
           this.customers = this.customers.filter(customer => customer.id !== customerId )
         },
         error : err => {
           alert("Erreur lors de la suppression !!!!")
+        }
+      })
+  }
+
+  searchCustomer(keyword: string) {
+    this.customerService.searchCustomer(keyword)
+      .subscribe({
+        next : data => {
+          this.customers = data
+        },
+        error : err => {
+          alert("Error During Search")
         }
       })
   }
